refactor(file-structure): create a common structure for modules
