+++ src/auth/auth.module.ts
@@ -1,9 +1,9 @@
 import { Module } from "@nestjs/common";
 import { DatabaseModule } from "src/database/database.module";
 import { AuthController } from "./auth.controller";
-import { AuthService } from "./auth.service";
-import { TokenService } from "./jwt.service";
-import { PasswordService } from "./password.service";
+import { AuthService } from "./services/auth.service";
+import { TokenService } from "./services/jwt.service";
+import { PasswordService } from "./services/password.service";
 
 @Module({
 	imports: [DatabaseModule],
