+++ src/database/services/connection.provider.ts
@@ -3,7 +3,7 @@ import { ConfigService } from "@nestjs/config";
 import { Kysely, PostgresDialect } from "kysely";
 import { Pool } from "pg";
 import type { AppConfig, DatabaseConfig } from "src/config/config.types";
-import type { DB } from "./generated.types";
+import type { DB } from "../generated.types";
 
 export const DB_CONNECTION = "CONNECTION";
